starter/requirements.txt
alembic==1.6.5
click==8.1.3
Flask==2.3.3
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.2
greenlet==1.1.0
gunicorn==20.1.0
Mako==1.1.4
psycopg2==2.9.9
python-dateutil==2.8.1
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.18
Werkzeug==2.3.7
astroid==2.2.5
ecdsa==0.13.2
future==0.17.1
isort==4.3.18
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.4.0
MarkupSafe==2.1.1
mccabe==0.6.1
# pycryptodome==3.3.1
pylint==2.3.1
python-jose==3.3.0
wrapt==1.11.1
python-dotenv==1.0.1